requirements/common.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pyodbc==4.0.30
Werkzeug==1.0.1
python-dotenv~=0.16.0
pytest~=6.2.2